add controllers 1.0

+++ src/app.js
@@ -1,15 +1,20 @@
 import express from 'express';
 import cors from 'cors';
 import * as todoService from './services/todo.services.js';
+import * as todoController from './controllers/todo.controller.js';
 const app = express();
 
-let todos = [
-  { id: '1', title: 'express', completed: true },
-  { id: '2', title: 'node', completed: true },
-  { id: '3', title: 'app', completed: false },
-];
-
 app.use(cors());
+
+app.get('/todos', cors(), todoController.get);
+
+app.get('/todos/:id', todoController.getOne);
+
+app.delete('/todos/:id', express.json(), todoController.remove);
+
+app.post('/todos', express.json(), todoController.create);
+
+app.put('/todos/:id', express.json(), todoController.update);
 
 app.patch('/todos', express.json(), (req, res) => {
   const search = req.query;
@@ -21,7 +26,11 @@ app.patch('/todos', express.json(), (req, res) => {
       return;
     }
 
-    todoService.deleteManyTodos(ids)
+    if (!ids.every((id) => todoService.getById(id))) {
+      throw new Error();
+    }
+
+    todoService.deleteManyTodos(ids);
 
     res.sendStatus(204);
     return;
@@ -32,74 +41,28 @@ app.patch('/todos', express.json(), (req, res) => {
       res.sendStatus(422);
       return;
     }
-    todoService.updateManyTodos(items)
 
-    res.sendStatus(204);
+    const errors = [];
+    const results = [];
+
+    items.forEach((item) => {
+      const { id, title, completed } = item;
+
+      const curentTodo = todoService.getById(id);
+
+      if (!curentTodo) {
+        errors.push({ id, title, completed, error: 'Not found' });
+      } else {
+        const result = todoService.update({ id, title, completed });
+        results.push(result);
+      }
+    });
+    res.status(200).json({ errors, results });
+
     return;
   }
 
   res.sendStatus(422);
-});
-
-app.get('/todos', (req, res) => {
-  res.send(todoService.getAll());
-});
-
-app.get('/todos/:id', (req, res) => {
-  const { id } = req.params;
-
-  const todo = todoService.getById(id);
-  if (!todo) {
-    res.sendStatus(404);
-    return;
-  }
-  res.send(todo);
-});
-
-app.delete('/todos/:id', express.json(), (req, res) => {
-  const { id } = req.params;
-
-  if (!todoService.getById(id)) {
-    res.sendStatus(404);
-    return;
-  }
-
-  todoService.deleteTodo(id);
-  
-  res.sendStatus(204);
-});
-
-app.post('/todos', express.json(), (req, res) => {
-  const title = req.body.title;
-
-  if (!title) {
-    res.sendStatus(422);
-    return;
-  }
-
-  const todo = todoService.create(title);
-  res.statusCode = 201;
-  res.send(todo);
-});
-
-app.put('/todos/:id', express.json(), (req, res) => {
-  const { id } = req.params;
-  const { title, completed } = req.body;
-
-  const todo = todoService.getById(id);
-  if (!todo) {
-    res.sendStatus(404);
-    return;
-  }
-
-  if (typeof title !== 'string' || typeof completed !== 'boolean') {
-    res.sendStatus(422);
-    return;
-  }
-
-  const updatetdTodo = todoService.update({ id, title, completed });
-
-  res.send(updatetdTodo);
 });
 
 app.listen(3005, () => {
